msm: clock: Add auto_off() function
Sometimes a clock is on upon boot and nobody is using the clock
so it should be turned off. We used to turn it off by calling to
proc_comm and not worrying about it. When we introduce clock_ops
we still just called clk_disable() and didn't worry about any
issues. Now that clocks are locally controlled there are 3 levels
of reference counts disturbed when the clk_disable() function is
used to turn off unused clocks.

Instead of using clk_disable() implement a new function
auto_off() to turn off clocks which are on at boot and nobody has
used. This helps the local clock implementation avoid messing up
the reference counts it maintains in its normal clk_disable path.

Change-Id: I57172186b217749536bdec666f461a06ac7596fa
Signed-off-by: Stephen Boyd <sboyd@quicinc.com>