msm_fb: Round LCDC pixel clock rate
The lcdc pixel clock rate varies among SoCs. Historically calling
clk_set_rate() on the pixel clock would send the request to
proc_comm wherein which the rate would be rounded to the nearest
frequency. Once we migrate to local clocks this will no longer be
the case and the rounding must be done prior to calling
clk_set_rate(). Therefore round the rate here (if the clock is
proc_comm controlled rounding is a nop).

Change-Id: I7be7a532b448707c997f215614b1c871e4d702ba
Signed-off-by: Kuogee Hsieh <khsieh@quicinc.com>