msm: audio: 7x30: Add proper error handling for failed adsp get scenario
Problem
The encoder drivers currently do not handle the error case properly if it
fails to get adsp resource. The system may eventually keep blocking the
resource for ever. This will impact the performance as will prevent any
suspend request.

Impact
The audpreproc task would be kept enable if the first session fails to get
the adsp resource. This would also initialize wake locks and would not free
them even if no encoder session is running.The system will never goto
suspend.

Solution
The driver is modified to disable the already enabled audpreproc task if it
fails to aquire the adsp task.

Change-Id: Ib96fdc8fb2cd743744dd940eb0ee89e20fd631c1
Signed-off-by: Asish Bhattacharya <asishb@qualcomm.com>