msm: audio: 7x30/8x55: Check 8x55 machine type for device registration
Audio driver registers different set of sound devices based on machine
type: SURF, FFA, or FLUID. Recent change to differentiate 7x30 and 8x55e
machine type caused sound devices for FLUID to be registered. Since
handset is not on FLUID device which is default Rx device for SURF. Audio
is found to be broken. Update audio driver to look for 8x55 machine type
as well

Change-Id: I986b0f9ed33a54109638b6eefa6636638584b65d
Signed-off-by: Patrick Lai <plai@quicinc.com>