msm_fb: Modified MDP3.1 workaround for artifacts
Source and Destination region matching is changed to reduce
artifacts in case rotation is 90, 180 and 270.
Older ROI were not placed correctly on framebuffer and were
producing artifacts.

In case dst_width mod 32 is 14, 22 and 30, equations to
split the destination ROI is changed to get even width/
height. This removes artifacts for YCbCrH2V1 color format.
as if the width is odd MDP makes it even in HW.

Added another case to break source and destination regions when
destination ROI width/height mod 32 is 2

MDP_CGC_EN register is set before PPP kick of and reset after
PPP complete. This sets the gating of PPP block and avoid
premature PPP interrupt for very thin ROI blits.

Change-Id: I80100c42430ea21f9a7562ff9e8f13dc16d90db3
Signed-off-by: Neeraj Kumar <c_neekum@quicinc.com>
Signed-off-by: Annamaria Pyreddy <c_apyred@quicinc.com>