msm: don't create useless page tables if there is no debug UART
If CONFIG_DEBUG_LL is set, the kernel attempts to map in the
UART so that debugging messages may be output.

MSM supports a choice of three UARTs to output on, but setting no
such UART is also supported. In that case, the kernel incorrectly
assumes that the UART is at 0 and maps in 512M of memory starting
at 0 for the non-existant UART.

Normally, this only creates extra mappings that are useless.
However, if there is memory below 0x20000000 on ARM v7, this
causes the page tables for this memory to be incorrectly set.

Change-Id: I660cb83b8701eb1f97f0ef2999cc71a639234931
CRs-fixed: 234233
Signed-off-by: Larry Bassel <lbassel@quicinc.com>