arm: perform a dmb in do_cache_op() on MSM 7x27
The ARM-specific cacheflush system call allows userspace to
flush the cache for a specified address range. The kernel
performs the cache flush on behalf of the user.

On MSM 7x27, this cache flush must also flush the AXI bus
buffer (or there may be data corruption), which can
be done in several ways, including performing a dmb().

Change-Id: I06f2ae853dab9be63f87f7abb8de96d53fecd6ad
CRs-Fixed: 236162
Signed-off-by: Larry Bassel <lbassel@codeaurora.org>