msm: 7x27: Setup Zero Page required for barriers and cacheops
A Zero page is required to simulate a write to a strongly ordered
memory location to flush the AXI bus buffer in 7x27. The various
barriers and cache operations depend on this page being available
when write_to_strongly_ordered_memory() is invoked. Since barriers
can be called from interrupt contexts this zero page should be initialized
prior to any of these calls.

Change-Id: I6b533810d391003f5cfc34fd6191609d8743ead4
Signed-off-by: Jessica Gonzalez <jgaona@codeaurora.org>