msm: audio: 7k: Reset the state during AUDIO_START only when disabled
Before sending enable command to DSP decoder the dec_state variable is
set to none state. The dec_state is set to success state on receiving
decoder enable event from DSP

This is being done even if AUDIO_START ioctl is called multiple times
when decoder is already in enabled state. In this case, the dec_state is
not set to success state as enable command is not sent to DSP again and
there will not be any decoder enable event from DSP. This will set the
decoder state variable to wrong state even when decoder is in enabled
state and will result in AUDIO_START failure

Hence moving the dec_state variable to none state during AUDIO_START
only if decoder is in disabled state

Change-Id: I15bff4b9cb46bf7ded277bc8cd16d30a58769ba7
CRs-Fixed: 245822
Signed-off-by: Sriranjan Srikantam <c_ssrika@qualcomm.com>