msm: 7x27: Fix msm_arch_idle() to preserve the state of I,A & F
msm_arch_idle() explicitly disables Aborts and Fiqs alone. This
can cause a state mismatch if the caller already has one of
these flags disabled because msm_arch_idle() enables them on
exiting which is incorrect. Make sure that we save the entire
state in CPSR and restore it correctly when exiting the function.

We additionally explicitly disable 'A', 'I' and 'F' so that we
don't take them within msm_arch_idle().

Change-Id: Ib503286d280f85ada0764be3026ec150c9b5d4e4
CRs-fixed: 253782
Signed-off-by: Naveen Ramaraj <nramaraj@codeaurora.org>
Signed-off-by: Vasantha Kumar <vasant@codeaurora.org>