msm: clock-pcom: Add pbus specific clock ops
The pbus clock always runs at half the ebi1 clock frequency.
Since ebi1 clock votes are dropped when the application processor
is power collapsed, the ebi1 clock is reserved for the acpuclock
driver. Other drivers which have ebi1 clock requirements that are
independent of application processor power status use the pbus.
To keep this change transparent across drivers and SoCs, add new
ops that divides the rate by 2 before sending the request to the
modem.

Change-Id: I87669536c0de01fb1163de6a35c7c5699f458d00
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>