USB: gadget: f_mass_storage: Fix USB WHQL Certification Issues
USB: gadget: f_mass_storage: Fix USB WHQL Certification Issues

As part of the USB MSCV testing, the host issues random commands and
expects the device to respond, but the device is not able to respond to
these commands, since the endpoints are not enabled due to the thread
latency and hence the compliance testing is failing.

Hence enabling/disabling the endpoints in interrupt context.

Change-Id: I27ad5875a3bde9a480866173d18cc3292c4af969
Signed-off-by: Chiranjeevi Velempati <cvelempa@codeaurora.org>