msm: Specify GPU pagetable params in the board file
Move the definitions for the desired GPU MMU virtual address range and
number of concurrent pagetables from the kgsl driver to the platform
definition where it belongs.  By their very nature, pagetables need to
be located in contigious memory and the kgsl driver will use this
information to allocate a large enough pool of contigious pages to
accomodate the requested parameters.

Change-Id: Ice6dc9d349324cd184c8f66b32a34732d73b2ddc
Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>