msm: timer: mask interrupts in set_mode
Previously the timer could wrap around and wake us up out of suspend,
since the timer interrupt has no interrupt enable bit to stop it from
triggering on the match value. So, explicitly mask and clear this
interrupt  when we suspend. There is no generic api for this, so call
irq_chip functions directly.

Change-Id: I0bd173ef2d5d70d611f74d060c15845259e85a19
Signed-off-by: Jeff Ohlstein <johlstei@codeaurora.org>
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>