GIC: move sysdev init to late_initcall
Inaccurate wakeup reasons are reported if gic_resume() is called before
gic_show_resume_irq(). This is because gic_resume switches from wake set
to the enabled set of interrupts.

So, the resume for sysdev in msm_show_resume_irq.c should be called before
the resume for sysdev in gic.c. Since resume callbacks are called
in the order of sysdev registration, register the gic sysdev in
late_initcall.

Change-Id: Idbb8549c94fe88a5a3d1f995d5ddb08f6f150047
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>