msm: timer: reset local timer count only at first boot
Previously we were resetting this timer each time we booted.
Sometimesalarms based on the old timer count were being used,
causing events to not come in until the timer caught up to the
old time. Fix this by notexplicitly resetting the timer except
on the first time we boot.

CRs-Fixed: 281314

Change-Id: I2ddc26a74ca5999be699b0fb2d7a12d0a9443626
Signed-off-by: Raj Kushwaha <rajk@codeaurora.org>