msm: subsystem-fatal-8x60: Wakeup the Q6SS after setting its NMI bit
To actually process the NMI, the Q6 needs to be awake, write
to a Q6 soft interrupt register to wake the q6 up.

This will allow the Q6 to process the NMI even when its
power collapsed, thus allowing it to flush caches and
allow logs to to be useful and coherent.

Signed-off-by: Vikram Mulukutla <markivx@codeaurora.org>
(cherry picked from commit e40cc8ad16477b28ef2d07df55ff6fd9a6dd4c21)

Change-Id: I4e9ce38c882ac663a3cb79ca497621e7a4f1b9f5
Signed-off-by: Shruthi Krishna <skrish@codeaurora.org>