msm: restart: Reset watchdog before re-enabling
Previously when we wanted to induce a restart, we simply disabled the
watchdog, set the bark and bite times, and re-enabled it. However, this
does not affect the watchdog count, and so sometimes it was already past
our bite time, and would not cause a reboot. So, reset the watchdog
count before re-enabling it.

Change-Id: If3450f01770f950ea086cb9b03807567c5ef9a8a
Signed-off-by: Raj Kushwaha <rajk@codeaurora.org>