msm_charger: Add support for alarm to signal charge start
Since we no longer hold a wakelock when charging is complete, we
need a mechanism to wakeup the system from suspend mode. The
battery alarm uses a wakeup interrupt to deliver such a
notification. Upon wakeup, the system starts charging again if a
valid charger is present. The voltage that's used for the battery
alarm wakeup is configurable through a module parameter.

Signed-off-by: Michael Bohan <mbohan@codeaurora.org>
(cherry-picked from 036d646803c269220ad094a8e8e2e07c40312fc7)

Change-Id: I3e15686ee214d10b88e6c23f33190aed42cb51a4
Signed-off-by: Michael Bohan <mbohan@codeaurora.org>