crypto: Fix corruption of dst buffer length
Fixes the corruption of the destination buffer lengths
when processing request of larger than 32 KB.

After a cipher operation (encryption or decryption) is complete,
data is copied to the destination buffer. For data packets larger
than 32 KB, data is processed in chunks of 32 KB. As a part of this
process, the destination buffer lengths are modified (to keep account
of how much data is already copied for each vbuf entry).

The destination buffer length for each vbuf entry, need to be
restored to its orginal value after the  operation is complete.

Signed-off-by: Mona Hossain <mhossain@codeaurora.org>
(cherry picked from commit 3a8c0e022eba819076cd13c99885303272105880)

Change-Id: I2933a7c973d7fa978989cfb5bcea842802a4838d
Signed-off-by: Shruthi Krishna <skrish@codeaurora.org>