msm: subsystem_restart: Provide additional info to callback functions
It may be useful to know what subsystem crashed when a
shutdown/powerup/ramdump callback is invoked. Provide this
information by passing in the name of the subsystem that
crashed as an argument to these callbacks.

Signed-off-by: Vikram Mulukutla <markivx@codeaurora.org>
(cherry picked from commit e045f68c44d5c3878cb6334f0af432418fc937ac)

Change-Id: Iac2f20c5328251157bd8cb1c62d9f5cc45b6078f
Signed-off-by: Raj Kushwaha <rajk@codeaurora.org>